server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
ultralytics==8.4.7
